pyproject.toml
[project]
name = "yelp-ai-assignment"
version = "0.1.0"
description = "Advanced AI Systems for Yelp Reviews"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.30.0",
    "datasets>=2.19.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "tabulate>=0.9.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]
